Review-based refactor
