+++ src/index.tsx
@@ -1,10 +1,9 @@
 import React from 'react';
 import { Provider } from 'react-redux';
-import { readEndpoint } from 'redux-json-api';
+
 import store from 'store';
 import { initializeDocumentHandlers } from 'init/documentHandlers';
 import { injectComponents } from 'init/components';
-import chromeStorage from 'chrome-storage';
 
 import './css/common/base.scss';
 // semantic-ui minimum defaults for semantic-ui to work
@@ -26,6 +25,7 @@ import PPSettings from 'PPSettings.interface';
 import * as chromeKeys from './chrome-storage/keys';
 import initializeChromeStorageHandlers from './init/chromeStorageHandlers';
 import { changeAppModes } from './store/appModes/actions';
+import { loadDataFromChromeStorage, loadInitialData } from './init/data';
 
 // Declared in webpack.config through DefinePlugin
 declare global {
@@ -33,30 +33,6 @@ declare global {
 }
 
 console.log('Przypis script working!');
-
-function loadInitialData() {
-  // This is our root request that needs to have part of the url (path) hardcoded
-  store.dispatch(readEndpoint('/annotations?url=' + window.location.href));
-}
-
-export function loadDataFromChromeStorage() {
-  return new Promise((resolve, reject) => {
-    chromeStorage.get([
-      chromeKeys.ANNOTATION_MODE_PAGES,
-      chromeKeys.DISABLED_EXTENSION,
-      chromeKeys.DISABLED_PAGES,
-    ], (result) => {
-      console.log(result);
-      const newAppModes = {
-        annotationModePages: result[chromeKeys.ANNOTATION_MODE_PAGES] || [],
-        isExtensionDisabled: result[chromeKeys.DISABLED_EXTENSION] || false,
-        disabledPages: result[chromeKeys.DISABLED_PAGES] || [],
-      };
-      store.dispatch(changeAppModes(newAppModes));
-      resolve();
-    });
-  });
-}
 
 /*
  * APPLICATION STATE POLICY
